Time: 27 ms (39.67%) | Memory: 50.9 MB (74.66%) - LeetSync

+++ 1408-find-the-smallest-divisor-given-a-threshold/find-the-smallest-divisor-given-a-threshold.java
@@ -1,12 +1,12 @@
 class Solution {
     static boolean isValidDivisor(int[] nums, int threshold, int divisor){
-        int sumOfResults = 0;
+        double sumOfResults = 0;
         double r = (double)divisor;
         for(int i = 0; i < nums.length; i++){
             double Quotient = Math.ceil(nums[i] / r);
             sumOfResults += Quotient;
         }
-        return sumOfResults <= threshold;
+        return (int)sumOfResults <= threshold;
     }
     public int smallestDivisor(int[] nums, int threshold) {
         int start = 1;
